requirements.txt
pytest
pytest-playwright
openpyxl
pytest-html